requirements.txt
django
djangerestframework
request